Scrum 5 player functions and Scrum 7 world basics (#1)
* Player Movement

Added in a temporary capsule for the player that has a script for player movements left and right.

* Player Interactions

Added a Player Interaction Script that allows the player to press 'E' and then prints to the console that they are interacting with something. This will be further developed once something interactable is added.

* Camera Following and Movement Fixes

A new script that follows the player with some delay for "smoothness" was added. Also some fixes with the movement where the player was too "floaty" because the y velocity was incorrectly set to 0 too often.