Imported Some Required UI Assets and Added a UI For Buiding
Made a UI that displays when the character is by a building plot and then it displays a number of buttons which will display building when developed further. For now a test building has been created to implement working functions for the menu. There was also some assets related to UI that Unity prompted me to import so those have been added as well.