Scrum 9 buildings system (#2)
* Added a System to Put Building Plots Into Game World

Instead of creating a grid system this will allow us to place "building plots" anywhere and the plots will handle what buildings can be placed there and also the position of the building.

* Imported Some Required UI Assets and Added a UI For Buiding

Made a UI that displays when the character is by a building plot and then it displays a number of buttons which will display building when developed further. For now a test building has been created to implement working functions for the menu. There was also some assets related to UI that Unity prompted me to import so those have been added as well.

* Added Three Basic Building Prefabs

Added the HAB, Solar Panel, and Research Lab as buildings you can put into the building plots.

* Last Minute Additions

These were last minute additions to clean up some confusing code and make the system work better. Mainly added code to make the adding of buildings to the plots simpler and also commented on all scripts.