Made GameManager and Interactions With Resources
Added some resources laid around the testing map. Added a GameManager script that helps track which object the player is trying to interact with and gives the player access to their Interact() function. Also made the buildings interactable for later design regarding showing UI stats for the buildings when interacted with.