Fixed Issues on Starting Screen
The start screen canvas was not sized properly or anchored right so that was fixed. Also the hierarchy of the game objects did not make sense so they were organized more clearly. Fixed an issue with the start screen settings exit button not appearing after closing settings screen once. Organized the folder system of scripts, prefabs, and materials better.