Scrum 48 technology tree basics (#5)
* Buttons and Title

Without  Background Image or Link to Game

* Re-add changes

I forgot to save earlier.

* Added Settings Screen and Exit Button

* Fixed X button of Settings Screen

* Added Basics of Technology Tree

Added UI elements for technology tree and much of the functions of it as well. We still need to design and completely finish functionality but that will be later after ideas for technology tree progression is complete.

* Fixed Issues on Starting Screen

The start screen canvas was not sized properly or anchored right so that was fixed. Also the hierarchy of the game objects did not make sense so they were organized more clearly. Fixed an issue with the start screen settings exit button not appearing after closing settings screen once. Organized the folder system of scripts, prefabs, and materials better.

---------

Co-authored-by: snsorl <suryasolaiappan@gmail.com>